+++ kiltergpt/__init__.py
@@ -2,5 +2,5 @@ from beartype.claw import beartype_this_package
 
 beartype_this_package()
 
-__version__ = "0.3.1"
-BEST_MODEL = "ilsenatorov/kilter-gpt/model-vvk6wo87:v0"
+__version__ = "0.4.0"
+BEST_MODEL = "ilsenatorov/kilter-gpt-finetune/model-tamkrd8d:v0"